requirement.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
idna==2.10
importlib-metadata==3.1.1
iniconfig==1.1.1
packaging==20.7
pluggy==0.13.1
py==1.9.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.2
requests==2.25.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
zipp==3.4.0
